+++ superwall/src/main/java/com/superwall/sdk/analytics/trigger_session/TriggerSessionManager.kt
@@ -1,23 +1,25 @@
 package com.superwall.sdk.analytics.trigger_session
 
-
-import android.app.Activity
+import com.superwall.sdk.Superwall
 import com.superwall.sdk.analytics.SessionEventsDelegate
 import com.superwall.sdk.analytics.SessionEventsManager
-import com.superwall.sdk.analytics.internal.TrackingResult
-import com.superwall.sdk.analytics.internal.trackable.Trackable
+import com.superwall.sdk.analytics.internal.track
+import com.superwall.sdk.analytics.internal.trackable.InternalSuperwallEvent
 import com.superwall.sdk.analytics.model.TriggerSession
-import com.superwall.sdk.analytics.session.AppSession
-import com.superwall.sdk.analytics.session.AppSessionManager
 import com.superwall.sdk.config.ConfigManager
+import com.superwall.sdk.config.models.getConfig
 import com.superwall.sdk.identity.IdentityManager
 import com.superwall.sdk.models.config.Config
-import com.superwall.sdk.models.paywall.Paywall
 import com.superwall.sdk.models.triggers.InternalTriggerResult
-import com.superwall.sdk.models.triggers.TriggerResult
 import com.superwall.sdk.paywall.presentation.internal.request.PresentationInfo
 import com.superwall.sdk.storage.Storage
-import com.superwall.sdk.store.abstractions.product.StoreProduct
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.flow.mapNotNull
+import kotlinx.coroutines.launch
+import kotlinx.coroutines.newSingleThreadContext
+import kotlinx.coroutines.withContext
+import java.util.UUID
 
 enum class LoadState {
     START,
@@ -25,119 +27,93 @@ enum class LoadState {
     FAIL
 }
 
+data class ActiveTriggerSession(val sessionId: String, val eventName: String)
 
 // TODO: https://linear.app/superwall/issue/SW-2366/[android]-implement-triggersessionmanager
 class TriggerSessionManager(
     val storage: Storage,
     val configManager: ConfigManager,
-    val appSessionManager: AppSessionManager,
     val identityManager: IdentityManager,
     val delegate: SessionEventsDelegate,
-    val sessionEventsManager: SessionEventsManager,
+    val sessionEventsManager: SessionEventsManager
+) {
+    private val singleThreadContext = newSingleThreadContext(name = "TriggerSessionManagerThread")
+    val pendingTriggerSessionIds: MutableMap<String, String?> = mutableMapOf()
 
-    ) {
-    val pendingTriggerSessions: MutableMap<String, TriggerSession> = mutableMapOf()
-    var activeTriggerSession: TriggerSession? = null
-    var observerCancellables: List<Any> = listOf()
-    var configListener: Any? = null
-    // TODO: Re-enable
-//    var transactionCount: TriggerSession.Transaction.Count? = null
-
-    fun listenForConfig() {
-        // implementation
+    /**
+     * The active trigger session tuple in format (sessionId, eventName)
+     */
+    suspend fun getActiveTriggerSession(): ActiveTriggerSession? = withContext(singleThreadContext) {
+        return@withContext activeTriggerSession
     }
 
-    fun addObservers() {
-        // implementation
+    private var activeTriggerSession: ActiveTriggerSession? = null
+
+    init {
+        CoroutineScope(singleThreadContext).launch {
+            listenForConfig()
+        }
+    }
+    private suspend fun listenForConfig() {
+        configManager.configState
+            .mapNotNull { it.getSuccess()?.getConfig() }
+            .collect { config ->
+                createSessions(config)
+            }
     }
 
-    fun willEnterForeground() {
-        // implementation
+    private fun createSessions(config: Config) {
+        // Loop through triggers and create a session ID for each.
+        for (trigger in config.triggers) {
+            pendingTriggerSessionIds[trigger.eventName] = UUID.randomUUID().toString()
+        }
     }
 
-    fun didEnterBackground() {
-        // implementation
-    }
-
-    fun createSessions(from: Config) {
-        // implementation
-    }
-
-    fun activateSession(
-        forPresentationInfo: PresentationInfo,
-        on: Activity? = null,
-        paywall: Paywall? = null,
+    suspend fun activateSession(
+        presentationInfo: PresentationInfo,
         triggerResult: InternalTriggerResult? = null,
-        trackEvent: (suspend (Trackable) -> TrackingResult)? = null
-    ) {
-        // implementation
+    ): String? = withContext(singleThreadContext) {
+        val eventName = presentationInfo.eventName ?: return@withContext null
+
+        val sessionId = pendingTriggerSessionIds[eventName] ?: return@withContext null
+
+        val outcome = TriggerSessionManagerLogic.outcome(
+            presentationInfo = presentationInfo,
+            triggerResult = triggerResult?.toPublicType()
+        ) ?: return@withContext null
+
+        activeTriggerSession = ActiveTriggerSession(
+            sessionId = sessionId,
+            eventName = eventName
+        )
+        pendingTriggerSessionIds[eventName] = null
+
+        triggerResult?.let {
+            val trackedEvent = InternalSuperwallEvent.TriggerFire(
+                triggerResult = it,
+                triggerName = eventName,
+                sessionEventsManager = sessionEventsManager
+            )
+            Superwall.instance.track(trackedEvent)
+        }
+
+        when (outcome) {
+            TriggerSession.PresentationOutcome.HOLDOUT,
+            TriggerSession.PresentationOutcome.NO_RULE_MATCH -> {
+                endSession()
+            }
+            TriggerSession.PresentationOutcome.PAYWALL -> {}
+        }
+
+        return@withContext sessionId
     }
 
-    fun endSession() {
-        // implementation
-    }
+    fun endSession() = CoroutineScope(singleThreadContext).launch {
+        val currentTriggerSession = activeTriggerSession ?: return@launch
 
-    fun enqueueCurrentTriggerSession() {
-        // implementation
-    }
+        // Recreate a pending trigger session
+        pendingTriggerSessionIds[currentTriggerSession.eventName] = UUID.randomUUID().toString()
 
-    fun enqueuePendingTriggerSessions() {
-        // implementation
-    }
-
-    fun updateAppSession(to: AppSession) {
-        // implementation
-    }
-
-    fun trackPaywallOpen() {
-        // implementation
-    }
-
-    fun trackPaywallClose() {
-        // implementation
-    }
-
-    fun trackWebviewLoad(forPaywallId: String, state: LoadState) {
-        // implementation
-    }
-
-    fun trackPaywallResponseLoad(forPaywallId: String?, state: LoadState) {
-        // implementation
-    }
-
-    fun trackProductsLoad(forPaywallId: String, state: LoadState) {
-        // implementation
-    }
-
-    fun trackBeginTransaction(of: StoreProduct) {
-        // implementation
-    }
-
-    fun trackTransactionError() {
-        // implementation
-    }
-
-    fun trackTransactionAbandon() {
-        // implementation
-    }
-
-    fun trackTransactionRestoration(withId: String?, product: StoreProduct?) {
-        // implementation
-    }
-
-    fun trackPendingTransaction() {
-        // implementation
-    }
-
-    fun trackTransactionSucceeded(
-        withId: String?,
-        forProduct: StoreProduct,
-        isFreeTrialAvailable: Boolean
-    ) {
-        // implementation
-    }
-
-    fun trackTransactionDeferred(withId: String?, forProduct: StoreProduct) {
-        // implementation
+        activeTriggerSession = null
     }
 }